This is a good one, so bear with me here...
We have a new HStoreThreadManager utility that we use to register EE and non-EE threads and then set their CPU core affinity.
Added new HStoreConf options site.cpu_affinity and site.cpu_affinity_one_partition_per_core that we can use to play around with different configurations on EC2 and understand what works the best. This provides a most speed boost per HStoreSite on EC2

Which leads me to now discover at the end of the day that what we really need is a separate incoming throttle per partition, not per HStoreSite. I've noticed some funkiness where one core will run slower than others on EC2. That might be because we keep queuing up incoming txns and have to acquire a lock in the ExecutionSite. Regardless this causes the HStoreSite to start shedding work, leaving some EE idle. We should really get rid of the single throttle and the ExecutionSite locks...
So that means I'm going to go home, eat dinner, and then bang out this fix.



git-svn-id: https://database.cs.brown.edu/svn/hstore/branches/partitioning-branch@2339 7a8f3d5b-8c5f-44cb-9125-2414a256df87